Refactorization: Updated remainder of the main function.
Poisson sampling section now utilizes the MLEM function rather than redoing the algorithm. Added functions to calculate dose and RMSD uncertainties in a vector (spectrum) or in a value (dose).

Variable names modified and arrays replaced with vectors for clarity.

Use of vectors is slower than arrays (have tested this out), but worth it for the improved legibility of the code. 1000 Poisson samples is still doable (<10 seconds in testing environment) and 50 samples achieves same result anyway (~1 second in testing environment).