Update Features: Several new user-definable parameters were added to the plot_spectra script. These include: the number of divisions on the y axis, line styling, etc. Specification of the color of spectra is now done using hexadecimal color codes, as in HTML, instead of ROOT colors to make it more compatible with other software (i.e. easier to copy color schemes).
Also updated the readSpectra function in the fileio.cpp source file (which is used by plot_spectra), to allow generation of plots per MU as well as plots per second. This is not surfaced to the user, but may be toggled by commenting/uncommenting the appropriate section in readSpectra AND updating the y-axis label in plot_spectra.