Feature update (unfinished): Update the "auto" unfolding technique such that it now iterates over a specified range (of iterations), all for a specified range of beta values. It calculates a parameter of interest at each (N,beta) combination (can toggle between total fluence, total dose, total "energy correction" from MAP, and maximum deviation in MLEM ratio). A 2D "matrix" of these POI values are ouput into a CSV file called map_file.csv.
Feature update (unfinished): Updated the runMAP() function to now have an option for the Median Root Prior in addition to the original quadratic prior term. This will be expanded and refactored to better accomodate additional priors in the future, and to surface the selection of a prior to the user.

Additional small updates.