New feature: Surface plots
Added a new plotting script to generate surface plots (contours). It can readily be used to plot a parameter of interest as a function of number of iterations and beta for the MAP method. The script reads in configuration data from a config file, similar to the earlier plotting scripts in this repository.