New feature: Uncertainty calculations based on J
Added the ability to generate spectrum uncertainty according to upper and lower J-thresholds as determined using a sigma_J parameter specified by the user. A custom class UncertaintyManagerJ was developed to handle the calculations and store relevant parameters for inclusion in the unfolding report. Uncertainties in the other calculated parameters, e.g. dose, accomodate this possibility as well. The user can specify which uncertainty type they wish to use using the uncertainty_type parameter.

Also made some adjustments to the root_helpers/plotSpectrum() method so that the output figure from unfolding is somewhat more visually appealing in accordance with recent modifications to the plot_spectra program. This will be useful for the upcoming project to develop a web interface for the unfolding program.