Miscellaneous post-paper-draft updates
New indicator function: J2. Also converges to 1 when within Poisson uncertainty, but uses the MD instead of the MSD in the numerator. This results in less substantial deviation from 1 when the criterion is not met. The denominator was empirically determined to be 0.798 * sqrt(mean) using MATLAB. The more complicated expression for MD of a Poisson distribution can numerically be shown to be equal to this for all mean values (small deviations at low mean values). Theoretically can sub in this expresseion in the MLEM-STOP method instead of J.

New plot_spectra settings:
- x_min and x_max: Use a new plotting approach to allow user to specify the upper and lower x bounds. By setting bounds at multiples of 10, can have upper and lower bounds indicated on x axis of the plot. Previously this was not the case. Use a "ghost histogram" approach to achieve this.
- y_digits_max: specify the maximum number of digits on y-axis tick labels before adopting scientific notation. This is the best and only way to achieve consistency on the y-axis scale for multiple related plots (e.g. multipart figure in the paper)
- error_fill_style: Allows user to set the ROOT fill style of the uncertainty region. This was necessary as the patterned fill options were ugly when outputting EPS or PDF plots. Should use the 1001 solid fill code when generating these plots.

Bug fix in plot_spectra: Previously the first plot was "sent to the back", resulting in the last plot being in the "front" of the plot. Now they are added in reverse order, so the first plot (usually the focal point of a figure) is at the front.

Bug fix in auto-generating a plot using plotSpectrum in unfold_spectrum. The figure path specified in the unfolding settings is now correctly applied when given.